♻ Refacto delete popup code
